refactor(dht, trackerless-network): Remove unused interfaces (#2149)
Removed interfaces which were not used in practice.

The removed interfaces had only one implementation, and the interfaces listed exactly exactly same methods as the public methods of implementing classes.  There was one exception: `IRouter` and `Router` had a mismatch in `doRouteMessage`. Because of that mismatch, the bug of https://github.com/streamr-dev/network/pull/2148 wasn't catch by the complier.

The removal helps us avoid that kind of mismatch bugs in the future. Also it helps us understand the code better, as we know exactly which class we are using in each of these cases. It also makes code navigation significantly more efficient as we can navigate directly to the implementation by clicking the method name.

Removed:
- `IRouter`
- `IRecursiveOperationManager`
- `IInspector`
- `IHandshaker`
- `INeighborFinder`
- `INeighborUpdateManager`

Some of these interfaces were used in `StrictRandomGraphNodeConfig`, but in practice we always provided the only implementations of those interfaces for the config. We don't currently plan to support user-pluggable interfaces, but if we want to support that in the future, we should remodel the public plugin API anyway (e.g. `INeighborUpdateManager` was just `{ start(): Promise<void>, stop(): void }`).

## Other changes

Moved `Methods` helper type from client's test utils to `test-utils` package.

## Future improvements

- Remove `IWebrtcConnection` if it is not needed?
- Use `mock` and `MockProxy` from `jest-extended` to avoid `as any` casting in the tests.
- Remove `ConnectorFacade` interface if/when we have only one implementation for that in the `dht` package.