fix(trackerless-network): Uniqueness check in `RandomGraphNode#getNeighborCandidatesFromLayer1` (#2209)
If we get two `PeerDescriptor` objects from `PeerManager`, can we assume that if the descriptors have the same `nodeId`, they are also the same object? Maybe we can't? If that is the case, this PR fixes the uniqueness check in `RandomGraphNode#getNeighborCandidatesFromLayer1`.

This solution is most likely slower than the previous solution.