feat(dht): Check version compatibility on handshakes (#2219)
## Summary

Check version compatibility on handshakes.

This ensures that nodes running incompatible older versions of the
software cannot join networks that assume the use of newer versions.

## Future Improvements

- The version field might no longer be needed in the handshake response
in the future. Once all nodes add the version number to the handshake
request we can safely assume that the requests version is enough to
decide compatibility.
- The version numbering could be separated from the monorepo version to
a protocol version. The Protocol version could be sent instead of the
monorepo version. This would ensure that backwards breaking changes to
the client or broker would not require network splitting on the lower
layers.
- The fallback value `BEFORE_TESTNET_TWO_VERSION` will likely not be
needed in the future and can be removed.

---------

Co-authored-by: Teo Gebhard <teo@streamr.com>