fix(dht): Recursive operation second path close condition (#2236)
## Summary

In cases where the second routing path for recursive operations was no
longer finding closer peers than the first path the
RecursiveOperationSession could easily enter timeout state. this is
because routing to the closest peer is not possible twice as it rejects
the second message as a duplicate. To counter this if the second path
finds the same closest node as the first path the operation will
resolve.

This change also speeds up closing RecursiveFindSession with the
FETCH_DATA mode when no data is behind the key

---------

Co-authored-by: Eric Andrews <eric.andrews@streamr.com>