feat(DHT): Router exclude parallel path root (#2235)
## Summary

The root of the primary routing path is excluded when routing on the
secondary path in recursive operations. This improves the creation of
the secondary path and fixes an problematic corner case in DHTs of 3
nodes in total.

The feature is only enabled for recursive operations as they are rarely
targeted towards exact nodeIds. Even in cases where they are, the exact
nodeId can be found without completing multiple routing paths. In the
remaining cases (forwarding, regular routing), the sent message is
targeted to a peer. In these cases it makes more sense to explore the
second path completely.

This feature fixes some of the flakyness seen in client and broker
end-to-end tests.

This is a breaking change headed towards testnet3