refactor(dht): Use nodeId in offering hash (#2256)
Refactor offerer analysis to use `DhtAddress` instead of `PeerIDKey`. The functionality doesn't change as both of these are the same hex representation of nodeId. This refactoring has been done so that we can remove `PeerID` class in the future.

## Future improvements

Improve test coverage related to the offerer analysis: e.g. `ConnectionManager` tests and `WebrctConnector` tests.