chore[!]: New signature method (#2286)
## Summary

Add new default signature type that is based on binary data. Retain
legacy support for Brubeck signatures.

## Changes
- Add mandatory `SignatureType` field to `StreamMessage` indicating
whether legacy (Brubeck) or current (1.0) signature method is used.
- Add unit tests ensuring that legacy signatures still pass signature
validation (see: legacySignatures.test.ts)
- Function `createSignaturePayload` now supports the two different
signature methods.

## Disclaimers

It should be noted that old stream messages from Brubeck cannot be used
as-is without converting the content and the signature fields. (Also the
serialization format has changed; we no longer support the array
serialization format but use protobuf serialization/deserialization
instead.)

The signature is a hex-encoded string in the old format, in the new
format it is a `Uint8Array` representing binary. This is a
straightforward conversion with `hexToBinary()`.

The content is a string in the old format, in the new format it is a
`Uint8Array` representing binary. This conversion is dependent on the
encryption type. If the content was encrypted, it was hex-encoded, so it
can be converted with `hexToBinary()`. If the content was not encrypted,
it was a string, so it can be converted with `utf8ToBinary()`.

Notice that the legacy signature format never coincides with binary
content type since that did not exist back then. Or simply put, whenever
we have legacy signature, we have JSON content type.

---------

Co-authored-by: juslesan <santeri.juslenius@gmail.com>
Co-authored-by: Henri Pihkala <henri@streamr.network>