refactor(protocol): rm runtime type validation from classes (#2299)
## Summary

Now that we don't de-serialize directly into `StreamMessage` but go thru
the protobuf equivalent first, there is (imo) no need to do runtime type
validation in the protocol class anymore. Semantic validation is still
reasonable to keep around (e.g. `msgId` is strictly ahead of
`prevMsgRef`, when using `AES` check that `groupKeyId` is present).

## Changes

- Remove `EncryptionType.RSA` (was only used internally, protobuf
equivalent has only NONE / AES)
- Remove Enum runtime validation in StreamMessage (ContentType,
MessageType, EncryptionType, SignatureType)
- Remove some runtime validation from `StreamMessage`, `MessageID`, and
`EncrypedGroupKey`.

## Limitations and future improvements

- We could add validation that if `LEGACY_SECP256K1` is used then
`BINARY` content type is disallowed.
- It is worth thinking thru where semantic validation should take place.
It is kind of late in the process to validate it only here in
`StreamMessage`, it could be argued that this should occur much earlier
in the process of de-serializing messages.