refactor(client): Use `urls` storage node metadata field (#2304)
Use `urls` field from storage node metadata to read the URL of the storage node. The field contains a `string` array. Previously the URL was stored in the `http` field and it contained only one URL instead of an array of URLs.

With this change we can configure different URLs for 1.0 clients and Brubeck era clients in the same contract.

From the url list we pick a random item. We don't currently support any retry logic within one storage node. (We already have a retry logic if there are multiple storage nodes.)

## Future improvements

- If there is no `urls` field in the metadata, the resend `Promise` rejects with an error `StreamrClientError`: Only absolute URLs are supported, code=`STORAGE_NODE_ERROR`. Maybe the error message could describe the root cause.
- We could maybe retry the resend by using another URL if a resend fails a randomly picked URL.