chore(ci): tweak release.sh and remove `prepublishOnly` NPM scripts (#2319)
## Summary

In bash script release.sh:
- Use `cd packages/<PKG> || exit` pattern. Safer to exit than to run
`npm publish` in an unintended dir.
- Build all packages in the very beginning (not strictly necessary if
following instructions, but maybe a good "just to be safe" thing to do,
it's a reasonably fast command anyway given that packages should be
built already at this point)
- Build client webpack first, then only start the release process.
(Kinda of sucks to have half of packages be published only to later
realize client webpack doesn't even compile. 🤦 That being said, this
shouldn't really happen if you are being patient and waiting for CI to
go green as the instructions recommend.)

NPM scripts:
- Remove `prepublishOnly` scripts from packages. (It's kind of a
safeguard, could be argued that it should be kept in. However we release
the monorepo as a whole, so you go thru the release.sh script anyway
ensuring everything is built. I would remove it here to make the release
process ever so slightly faster.)
- Remove client's `build-production`. (Enough to build webpack since the
node version will be built already, no point in rebuilding).