feat(network): Connectivity check versioning (#2320)
## Summary

Add a version field to the ConnectivityResponse. The server adds its
version to the response and the client will decide weather or not to
discard the response based on the received value.

## Other changes

Pumped up the number of failed connectivity requests before the stopping
connectivity checking from 5 to 10

## Limitations and future improvements

- When running connectivity checks on entry points, the entry points
that could not be reached should be tracked and excluded.
- Could also check for version mismatches on the server side. For now
older clients making connectivity requests to newer servers should be
compatible.
- The version field could be added to the PeerDescriptor. This way
incompatible connections can be avoided before ever opening a
connection.