feat(DHT): connectivity check even if websocket server was not started (#2348)
## Summary

Run connectivity check to discover own IP address even when WS server
was not started. Fixes a clustering around NodeIds starting with 85 for
all nodes without a WS server as they use ip `127.0.0.1`.

Configuring entry points is now necessary when running a DhtNode without
a WS server, unless the node is configured with the proper values to
generate a PeerDescriptor without a ConnectivityCheck.

(Is backwards compatible)

## Future improvements

- Could refactor ConnectivityRequestHandler further
- Change ConnectivityRequest port to an optional field
- Perhaps ConnectivityCheck could be renamed to reflect that any kind of
checking is no longer done. The point is to find out the node's
connectivity information
- Could rename the entry point config in the connection layer? Entry
point still makes since since ConnectivityChecks are an operation that
all entry points should support.