fix(DHT): clean up hanging invalid connections from the ConnectionManagers state (#2362)
## Summary

WebSocket client connections can sometimes remain hanging in the
connection managers state if the ConnectionManager acting as a client is
not running a WebSocket server itself. The client connections appear to
be targeted to the same WS server with a different nodeId. Something
could be going wrong with the handshake that should be rejecting new
connections if the nodeId does not mtach the server side.

Implementing this patch solution as the problem is rare and is thus
difficult to debug. The hanging connections don't seem to hinder the
nodes operation by that much but it is best to clean them up to avoid
memory leaks.