fix(network): fix infinite loop corner case in parallel handshakes (#2365)
## Summary

There was an unfortunate corner case where in some cases when trying to
find random peers to connect to a node could enter an infinite loop. It
is not clear if nodes ever entered it in production or testing but this
change ensures that the case is no longer possible.