feat(trackerless-network): Get local info via `fetchNodeInfo` (#2373)
## Summary

The `NetworkNode#fetchNodeInfo` can be called by any peer descriptor. Previously if it was called with the peer descriptor of the local node, the query failed with `CANNOT_CONNECT_TO_SELF` error.

Also changed the return type to be `NodeInfo` instead of `NodeInfoResponse`.
- The `NodeInfoResponse` is a low level RPC communication interface, and it is better to not to expose it in the `trackerless-network`'s public API.
- Both interfaces have the same fields, but in `NodeInfo` all the fields are required. (In practice all fields of `NodeInfoResponse` also also required, but we can't annotate the fields as such before we implement the protobuf message validation.)

## Refactoring

Extracted the `createNodeInfo` business logic from the RPC server to `NetworkStack`.