refactor(trackerless-network): `StreamMessage` structure (#2380)
## Summary

Divided `StreamMessage` entity to three sub entities: `ContentMessage`, `GroupKeyRequest` and `GroupKeyResponse`:

```proto
message StreamMessage {
  MessageID messageId = 1;
  optional MessageRef previousMessageRef = 2;
  bytes signature = 3;
  SignatureType signatureType = 4;
  oneof body {
    ContentMessage contentMessage = 5;
    GroupKeyRequest groupKeyRequest = 6;
    GroupKeyResponse groupKeyResponse = 7;
  }
}
```

The similar structure was already used in the `DhtRpc.proto`'s `Message` entity.

## Other changes

Small change in broadcast metrics: we now include only stream messages, not group key messages. This makes sense, as we already reported just the message content length, i.e. we didn't include e.g. the message headers or other kind of control data. The reason for this change is that there is no longer `content` field in group key messages from which we could calculate the content length.

## Open questions

Should the `previousMessageRef` field be in the`StreamMessage` entity or in the `ContentMessage`? Only content messages have references to previous messages, but maybe some messages in the future could also have that kind of sequencing. If we move it, it would make sense to move the `messageId#messageChainId` too as the previous message reference is always `messageChain`-specific. Therefore it would maybe be most consistent if we keep it in the `StreamMessage` (i.e. as it is now in this PR).

## Future improvements

The `DhtRpc.proto`'s `Message` entity has a separate `messageType` field. That is maybe not needed as we can infer the type from the `oneof` object (e.g. `if (msg.body.oneofKind === 'contentMessage') ...`).