refactor(trackerless-network, client): Group key binary signature (#2381)
## Summary

Changed the `StreamMessage#signature` payload for group key request and responses. Previously it was a  UTF-8 byte sequence of a custom JSON string. Now it is a compact byte sequence similarly to the `ContentMessage`.

## Refactoring

Changed the internal representation of `StreamMessage.content` for `GroupKeyRequest`/`GroupKeyResponse` objects. It is now the binary representation of the protobuf `GroupKeyRequest`/`GroupKeyResponse` entity. This change doesn't affect the end-users as group key requests/responses are never exposed from the client.

Moved the `signature.ts` from `protocol` package to the `client` package.

## Future improvements

Could simplify the signature payload generation slightly by re-ordering the `prevMsgRef` to be before the content (in that case there is always a generic header part, and other parts are message type specific).