chore(DHT): use WS library instead of websocket in the server side (#2383)
## Summary

Start using the ws library instead of websocket in the server side. This
is slight performance improvement when the number of received incoming
and outgoing messages to a node is high.

## Other changes

- Renamed ServerWebsocket -> WebsocketServerConnection.
- Removed `getRemoteAddress` and `getRemoteIP` from
WebsocketServerConnection. Instead the class has a new public field
`remoteIpAddress`. This is always the public facing IP address.
- Removed `NodeJsWsServer` variable from WebsocketServer as it is no
longer required for electron tests.

## Future improvements
- Disconnect WebsocketServerConnections if they receive string data

---------

Co-authored-by: Teo Gebhard <teo@streamr.com>