feat(network): Select layer2 neighbors with bias towards Hybrid DHT neighbors  (#2399)
## Summary

Use the Hybrid DHT in Layer1 to discover peers closer to the nodes
region.

Layer2 nodes will always prioritise opening new connections to the
closest left and right nodes in the ring. Nodes will not attempt to
actively connect to the closest found node in the ring if it is found,
while they have N or more connections. The nodes will simply prioritise
the closest left and right connection whenever a new connection is
required.

With this change nodes will have more layer2 neighbors with a priority
for geographical proximity. This should improve the mean message
propagation times in the streams.

## Future Improvements?

- Add slots for geographically nearby connections. This way if new
nearby contacts are found nodes will replace the existing out dated
ones. (Might not be good for churn tolerance)
- Start using `getRandomRegion` from the `cdn-location` library in local
simulation testing. (The latency table needs to be updated accordingly)
- Refactor the parallel handshake logic 
- Add a new object that contains all different contact types in Layer2
- Add more testing specifically for selecting neighbprs from the ring
and other options.

---------

Co-authored-by: Petri Savolainen <petri.savolainen@spaceify.net>