feat(dht): [NET-1256] Protocol version (#2404)
Send protocol version in handshake messages and in connectivity response. 

When two nodes negotiate whether they are compatible or not, it is up to the newer version to decide if it supports the old version or not.

The older version assumes optimistically that it may be supported by the newer version. It can't know for sure, but the other node will tell if it is not supported (e.g. rejecting the handshake with `UNSUPPORTED_VERSION` error).

## Future improvements

Could maybe add version info also to `ConnectivityRequest`