updated npm and docker image versions to reflect 1.0 release (#2422)
## Summary

Updated sdk & streamr dnoe version references to release.

## Changes

Provide a bullet list of individual changes. Leave this section out if
change
set is small and obvious from summary.

## Limitations and future improvements

Provide a bullet list or description of known omissions or limitations
of the
solution and/or any ideas for future improvement. Leave section ouf if
not applicable.

## Checklist before requesting a review

- [ ] Is this a breaking change? If it is, be clear in summary.
- [ ] Read through code myself one more time.
- [ ] Make sure any and all `TODO` comments left behind are meant to be
left in.
- [ ] Has reasonable passing test coverage?
- [ ] Updated changelog if applicable.
- [ ] Updated documentation if applicable.