refactor(trackerless-network): Rename `StreamrNode` (#2140)
Renamed `StreamrNode` class to `ContentDeliveryManager`. The instance is not a node in any of the networks. Instead it orchestrates the stream part delivery layer operations by managing `StreamPartDelivery` instances (each of which is a node in layer 2 or a proxy client).

## Other changes

Moved and rephrased the startup logging "Starting new StreamrNode..." from `StreamrNode` to `NetworkStack`.

## Future improvements

- Would it make sense if `NetworkStack` would provide the `createLayer1Node()` and/or `createRandomGraphNode()` functions?
- We could unify the naming in `NetworkOptions`? The fields and interfaces don't seem match. Actually after this PR there is negative correlation as the type of `layer0` is a `...Node`, and the type of `networkNode` is `...Manager`. (Also if there is no difference between `Options` and `Config`, we could maybe use only either of those). 
```ts
export interface NetworkOptions {
    layer0?: DhtNodeOptions
    networkNode?: ContentDeliveryManagerConfig
    ...
}
```