refactor(dht): Use `ConnectionLocker` instead of `ConnectionManager` (#2440)
Why `handleIncomingMessage` is no longer needed in `DhtNode`?  There is no need to pass messages to `ConnectionManager` as none of the message is for `ConnectionManager`'s service ID. 

PR #2375 already fixed problems related to calling `ConnectionManager#handleIncomingMessage` from the `DhtNode`. In this PR the method was completely removed as it was observed that the call is completely redundant.