refactor(dht): Remove unnecessary `SortedContactList` checks (#2443)
When there is an item in `contactIds` there must always be a corresponding key in the `contactsById` map. Therefore we don't need tho do the `if (contact)` checks.

Also refactored `getClosestContacts` to use `getAllContacts`.

## Future improvements

In some places where `SortedContactList#getAllContacts` is called, the caller should actually call `getClosestContacts` as the result is expected to be in the sorted order. Maybe we should remove `getAllContacts`? Or optimize it by so that it just returns the keys of `contactsById`?