fix(DHT), react to disconnected, failed and closed events in browser webrtc (#2427)
## Summary

Add a function for handling stateChanges in webrtc PeerConnection. If
the PeerConnection state is switched to disconnected, closed or failed
close the BrowserWebrtcConnection.

This helps in clenaing up connections after lost conenctivity in the
browser.

## Future improvements

Investigate if a similar fix is possible to add to the websocket client
connections or is ping/pong the only option.