feat: Topology visualization tool (#2478)
Add new developer tool for topology visualization. It reads a network topology definition from a JSON and creates a DOT graph. The JSON may also contain labels for some of the nodes and a color-coded route information.

## Future improvements

Now the tool is in new `developer-tools` directory to monorepo root. But maybe we could implement this as a `cli-tools` command.