test(dht): `DhtNode` integration test (#2480)
Added a simple integration test for `DhtNode`. In the test there is an entry point, and 3 other nodes using the same _fake environment_, i.e. transports which propagate messages between instances just by calling `transport.send()` methods. The entry point and other nodes are minimalistic `DhtNodeRpcLocal` instances.

In the test the local joins a DHT. Soon after the join it should have 4 neighbors in the network (the entry point and the 3 other nodes).

Fixed also a `this` binding issue in `DhtNode`.

## Future improvements

- The connection management in `FakeTransport` is currently very straighforward. Maybe the connections could be bidirectional: i.e. if there is a connection from `A` to `B`, there should also be a connection from `B` to `A`. Maybe it could also have some disconnection logic or max connection count?