feat(dht): Optimize `DhtNode#getClosestContacts` (#2481)
Before this PR `DhtNode#getClosestContacts` did unnecessary sorting. The `PeerManager#closestContacts `is always sorted in reference to the node's ID, and therefore we can just return the nodes in that order from `DhtNode#getClosestContacts`.

Moved the `getClosestContactsTo` functionality from `PeerManager` to a separate helper function. The only place that functionality is currently needed is the `DiscoverySession#findContacts`. There we query the `PeerManager`'s contacts and pass those to the `getClosestContacts` helper function. The helper function receives the contact list as an `Iterable`, and therefore it can take inputs also from other sources in the future.

As `PeerManager#getClosestContacts` now returns a `PeerManager`'s data without copying it, the return type has changed to `ReadonlySortedContactList<DhtNodeRpcRemote>`. That way we can ensure that other components can't corrupt `PeerManager`'s internal state by modifying the object. 

## Benchmark

- 200 contacts in the `PeerManager`
- do 1000 queries

`DhtNode#getClosestContacts`:

- before: ~3400 ms
- after: ~10 ms

## Future improvements

- Other `PeerManager` methods (e.g. `getClosestNeighborsTo`) could use this same pattern
- Maybe we could use `getClosestContacts` helper method in some places where we now create a temporary `SortedContactList`  instance just to order the contacts (maybe bit more readable, doesn't affect performance)