feat(DHT): Ping neighbors and ring contacts in layer1 (#2501)
## Summary

Added configuration to periodically ping neighbors and ring contacts.
This ensures that the peers in the nodes' structures are periodically
refreshed. This ensures that ungracefully leaving peers are eventually
detected.

## Changes

- Added utility function to ping contacts.
- Added two new configurations to the DhtNode to enable periodic
pinging.
- Query random neighbors in the PeerDiscovery's Recovery interval

## Limitations

There needs to be new mechanisms implemented to ensure that nodes that
lose connectivity to the internet can rejoin the layer1 DHT. Previously
the contacts and nodes in the buckets would remain there if internet
connectivity was lost. Nodes will no longer be able to rejoin the
streams properly. Publishing this feature before such mechanisms are in
place could be detrimental to the system in the long run.