refactor(dht): Use `getClosestNodes` for neighbors (#2529)
Use `getClosestNodes` helper function to analyze closest nodes instead
of querying the sorted data from `PeerManager`. This is a follow-up
refactoring to a PR #2481.

## Changes

- Removed `PeerManager#getClosestNeighborsTo`
- Changed `PeerManager#getNeighbors` return type
- it is now similar to two other node accessors (`getNearbyContacts`,
`getRandomContacts`) which also return a list of `DhtNodeRpcRemote`
instances
- Refactored `getClosestContacts` utility function to operate on
`PeerDescriptor` instances, and renamed it to `getClosestNodes`

## Open questions

Is there any need to create a new `DhtNodeRpcRemote` in `DhtNode:316,`
or could the pruning use the instance returned by
`peerManager.getNeighbors()`? How about other places where
`createDhtNodeRpcRemote` is called: would it be ok to do a lookup from
`PeerManager` instead?

## Future improvements

- Migrate all uses of `getClosestContacts` to use `getClosestNodes`
instead
- Maybe could simplify/optimize the analysis of `StoreManager`'s
`selfIsPrimaryStorer` (currently it resorts the node list)