refactor(dht): Optimize storer analysis (#2533)
Optimized storer analysis in `StoreManager`. Before this PR we sorted the candidates twice: first by querying `getClosestNodes(key, this.config.getNeighbors(), ...),` and then by adding those contacts to a new `SortedContactList`. Now we sort the list only once.

Also refactored the class. And made test cases more realistic by including the added contact in the neighbors list (as it would be in this kind of small network).

## Benchmark 1
- 50 neighbors
- do 1000 `onContactAdded` calls, i.e. benchmark the `replicateAndUpdateStaleState` method
- before: ~1200 ms
- after: ~1000 ms

## Benchmark 2
- 50 neighbors
- do 1000 `replicateDataToNeighbors` calls
- before: ~520 ms
- after: ~370 ms

## Future improvements

- Improve test coverage by adding a test in which the new contact is not a neighbor (i.e. it is some other nearby contact).