fix(DHT): fix hanging connection issue with connectivity requests (#2519)
## Summary

When ws connections are opened using connection requests, the server
checks if the requested connection's presumed peer descriptor is
correct. This fixes a problem with hanging connections detected in nodes
that do run a WS server.

## Future improvements

This problem went undetected due to the case not being tested. The
connection layer needs thorough refactoring to make testing possible.