chore(client) [NET-1287]: use `FallbackProvider` (#2562)
## Summary

Use `FallbackProvider` to provide redundancy when a RPC is down. 

Also use custom event listener implementation (the `ChainEventPoller`
class) as the `FallbackProvider` would poll too many RPCs.

## Refactoring

- Renamed `RpcProviderFactory` to `RpcProviderSource`
- Moved `contracts.ts`

## Future improvements

- Maybe the custom event listener implementation is no longer needed
when `FallbackProvider` implements `eventQuorum` and `eventWorkers`
functionality. Both of these options are already available, but there is
a code comment which states that the functionality is "currently
unimplemented":
https://github.com/ethers-io/ethers.js/blob/92761872198cf6c9334570da3d110bca2bafa641/src.ts/providers/provider-fallback.ts#L171
- What would be a good value for `pollInterval`. Previously we didn't
define an explicit value in the client, and therefore 4s interval was
used?
https://github.com/ethers-io/ethers.js/blob/92761872198cf6c9334570da3d110bca2bafa641/src.ts/providers/abstract-provider.ts#L428
---------

Co-authored-by: Teo Gebhard <teo@streamr.com>