feat(client, broker): [NET-1310] Use custom event polling logic for Broker features (#2601)
Moved `OperatorContractFacade` and `operatorContractUtils` from the
`broker` package to the `client` package. Renamed it to `Operator`.
Added `StreamrClient#getOperator` semi-public method to the client API.

Modified the class to use client's contract utilities, e.g. the
`FallbackProvider` and the `ChainEventPoller`.
- The functions of `operatorContractUtils` don't use `FallbackProvider`
and the `ChainEventPoller`. That's maybe ok as those are used in tests
only. But if we start to use the functions in production code, we should
start to start to use the helper classes also in these functions.

## Future improvements
- Replace individual contract artifacts with imports from
`@streamr/network-contracts-ethers6`. But need to find a way to do this
without bloating webpack bundle size.
- Could use https://typedoc.org/tags/hidden/ in classes/methods which
are currently marked as `@deprecated`?
- Do we need to export the `_operatorContractUtils` object? Maybe could
add a `StreamrClient#getSponsorship` method? Some methods/functions of
`Operator` and `operatorContractUtils` could be moved to the class
returned by that method
  - Maybe could remove `OperatorRegistry` class? 
- Lazy load instances of `Contract` in `Operator`?

---------

Co-authored-by: Eric Andrews <harbu88@hotmail.com>