fix(sdk): Fix `Operator#voteOnFlag` gas estimation (#2734)
We call the `estimateGas` function via the decorated contract. Before
the PR the decorated contract didn't have any of these
`BaseContractMethod` implemented. Therefore a call to `voteOnFlag`
always failed with error
```
this.contract.voteOnFlag.estimateGas is not a function
```

Added the method to the decorated contract. Also modified to the test
case to assert the expected error.