fix(sdk): Flag expiration timestamp (#2739)
Before this PR the expiration timestamp is now calculated by adding
`maxAge` to `flaggingTimestamp`. Now `voteEndTimestamp` is used instead
of `flaggingTimestamp`. It ensures that the flag can't expire before the
voting ends.

## Behavior before the fix

If a small value was configured for `maxAge`, the flag expired before
the voting ended. In that case the `closeExpiredFlags` service tried to
close the "expired" flag during the voting time. That caused the
operator to cast a `NO_KICK` vote instead of closing the flag.

This is not an issue in the production environment, as the default value
for `maxAge` is one day.