feat(node): Do not run `checkOperatorValueBreach` immediately (#2753)
## Changes

Changed the `executeAtStart` parameter of `checkOperatorValueBreach`
scheduler. Now the task is not run immediately after the plugin starts.
Instead it is first run after one hour. This is maybe a good change as
`checkOperatorValueBreach` is not a high priority task.

## Style

Moved the code block to be the last scheduled operation. This doesn't
the functionality, but is more readable:
- first we schedule important tasks which are run immediately
(`announceNodeToStream`, `announceNodeToContract`,
`maintainOperatorValue`)
- then lower priority task which are not run immediately
(`inspectRandomNode`, `closeExpiredFlags`, `checkOperatorValueBreach`)